bpf: Avoid implicit shorten-64-to-32 in clang 19
As per the below PR, clang 19.1.x is having -Wshorten-64-to32 under the
-Wimplicit-int-conversion, which is already forbidden in Makefile.bpf.
This commit is to avoid the conversion error as per below sample

```
./lib/nodeport.h:1825:18: error: implicit conversion loses integer precision: '__u64' (aka 'unsigned long long') to '__s32' (aka 'int') [-Werror,-Wshorten-64-to-32]
 1825 |         __s32 len_old = ctx_full_len(ctx);
```

Relates: https://github.com/llvm/llvm-project/pull/80814

+++ bpf/tests/xdp_nodeport_lb4_test.c
@@ -108,7 +108,7 @@ static __always_inline int build_packet(struct __ctx_buff *ctx)
 	data += sizeof(struct tcphdr) + sizeof(tcp_data);
 
 	/* Shrink ctx to the exact size we used */
-	offset = (long)data - (long)ctx->data_end;
+	offset = (int)((long)data - (long)ctx->data_end);
 	bpf_xdp_adjust_tail(ctx, offset);
 
 	return 0;
